packages/shorewall-nft-stagelab/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shorewall-nft-stagelab"
version = "1.8.0"
description = "Distributed bridge-lab for shorewall-nft performance and readiness testing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL-2.0-only"}
authors = [
    {name = "André Valentin", email = "avalentin@marcant.net"},
]
keywords = [
    "shorewall", "nftables", "firewall", "testing", "performance", "netns", "bridge",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Networking :: Firewalls",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "shorewall-nft-netkit>=1.8.0",
    "shorewall-nft>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[project.scripts]
shorewall-nft-stagelab = "shorewall_nft_stagelab.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["shorewall_nft_stagelab*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E701", "E731", "E741", "F601", "F841"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]
